support for MSP fields ActualWork ActualDuration
changed these fields' type from Integer to Float (in MSP)
to be consistent with other connectors and fields.

added loading these fields from MSP file. the current MPXJ library
always returns 0.0 for these fields in an MSP-like XML file
produced by ProjectLibre. Need to try a real MSP 2013 file
and new version of MPXJ library.

+++ connectors/connector-msp/src/main/scala/com/taskadapter/connector/msp/GTaskToMSP.scala
@@ -71,8 +71,8 @@ class GTaskToMSP(mspTask: Task, resourceManager: ResourceManager) {
 
       case MspField.taskDuration.name => mspTask.setDuration(Duration.getInstance(value.asInstanceOf[Float], TimeUnit.HOURS))
       case MspField.taskWork.name => mspTask.setWork(Duration.getInstance(value.asInstanceOf[Float], TimeUnit.HOURS))
-      case MspField.actualWork.name => mspTask.setActualWork(Duration.getInstance(value.asInstanceOf[Int], TimeUnit.HOURS))
-      case MspField.actualDuration.name => mspTask.setActualDuration(Duration.getInstance(value.asInstanceOf[Int], TimeUnit.HOURS))
+      case MspField.actualWork.name => mspTask.setActualWork(Duration.getInstance(value.asInstanceOf[Float], TimeUnit.HOURS))
+      case MspField.actualDuration.name => mspTask.setActualDuration(Duration.getInstance(value.asInstanceOf[Float], TimeUnit.HOURS))
       case MspField.percentageComplete.name => mspTask.setPercentageComplete(value.asInstanceOf[Int])
       case MspField.actualFinish.name => mspTask.setActualFinish(value.asInstanceOf[Date])
